fix(desktop): restore PostHog identity stitching for pre-sign-in events
Anonymous pre-sign-in events (desktop_opened, /sign-in pageview) weren't
merging into the user they later identify as, blinding the top of the
activation funnel — stitching crashed ~77% -> ~4% the week #4858 shipped.

Two causes:
1. #4858 bootstrapped posthog-js with the machine id as the anonymous
   distinctID; a stable shared device id doesn't follow the normal
   anon->identified merge on identify().
2. reset() ran reactively on every signed-out session state (incl. the
   initial anonymous load and token rotation), rotating the anon id and
   orphaning pre-sign-in events.

- Remove the device-id bootstrap (posthog.ts, PostHogProvider.tsx) — back
  to posthog's persistent localStorage anon id, which merges correctly.
- Move reset() + analytics cleanup into a useSignOut() hook wired to the
  real user-initiated sign-out buttons only (AccountSettings,
  OrganizationDropdown, create-organization) — never the AuthProvider
  token-rotation path.
- Simplify PostHogUserIdentifier to the canonical keyed identify effect
  (no refs), per PostHog's React guide.
