+++ apps/desktop/src/renderer/routes/_authenticated/settings/account/components/AccountSettings/AccountSettings.tsx
@@ -4,6 +4,7 @@ import { Input } from "@superset/ui/input";
 import { toast } from "@superset/ui/sonner";
 import { useLiveQuery } from "@tanstack/react-db";
 import { useEffect, useState } from "react";
+import { useSignOut } from "renderer/hooks/useSignOut";
 import { apiTrpcClient } from "renderer/lib/api-trpc-client";
 import { authClient } from "renderer/lib/auth-client";
 import { electronTrpc } from "renderer/lib/electron-trpc";
@@ -47,9 +48,7 @@ export function AccountSettings({ visibleItems }: AccountSettingsProps) {
 
 	const user = usersData?.find((u) => u.id === currentUserId);
 
-	const signOutMutation = electronTrpc.auth.signOut.useMutation({
-		onSuccess: () => toast.success("Signed out"),
-	});
+	const signOut = useSignOut();
 
 	const selectImageMutation = electronTrpc.window.selectImageFile.useMutation();
 
@@ -162,7 +161,10 @@ export function AccountSettings({ visibleItems }: AccountSettingsProps) {
 						>
 							<Button
 								variant="outline"
-								onClick={() => signOutMutation.mutate()}
+								onClick={async () => {
+									await signOut();
+									toast.success("Signed out");
+								}}
 							>
 								Sign out
 							</Button>
